AFS creation and extraction rewrite.
It's easier for the user to use. The user no longer has to try different command options in order to create appropriate AFS files for each different game. When extracting an AFS file, the program will generate a JSON file with metadata of how the AFS file is built. That will be used again when regenerating it to make sure it preserves the same format. If no metadata is present when creating an AFS file, it will be created with default settings.